Accept contacts refactor

+++ lib/dao/wish_list_dao.dart
@@ -15,7 +15,11 @@ class WishlistDao {
   final FirebaseFirestore _db = FirebaseFirestore.instance;
 
   Stream<QuerySnapshot<Map<String, dynamic>>> getWishlistsStreamSnapshot(String userId) {
-    return _db.collection('wishlists').where('ownerId', isEqualTo: userId).snapshots();
+    return _db
+      .collection('users')
+      .doc(userId)
+      .collection('wishlists')
+      .snapshots();
   }
 
   Stream<QuerySnapshot<Map<String, dynamic>>> getSharedWishlistsStreamSnapshot(String userId) {
@@ -23,24 +27,39 @@ class WishlistDao {
     if (currentUser == null) {
       throw Exception('Usuario no autenticado.');
     }
-    return _db.collection('wishlists').where(
-      Filter.and(
-        Filter.or(
-          Filter('sharedWithContactIds', arrayContains: currentUser.uid),
-          Filter('privacy', isEqualTo: 'public')
-        ), 
-        Filter('ownerId', isEqualTo: userId)
-      )
-    ).snapshots();
+    return _db
+      .collection('users')
+      .doc(userId)
+      .collection('wishlists')
+      .where(
+        Filter.and(
+          Filter.or(
+            Filter('sharedWithContactIds', arrayContains: currentUser.uid),
+            Filter('privacy', isEqualTo: 'public')
+          ), 
+          Filter('ownerId', isEqualTo: userId)
+        )
+      ).snapshots();
   }
 
   Future<QuerySnapshot<Map<String, dynamic>>> getWishlistsStream(String userId) {
-    return _db.collection('wishlists').where('ownerId', isEqualTo: userId).get();
+    return _db
+      .collection('users')
+      .doc(userId)
+      .collection('wishlists').get();
   }
 
   Future<String> createWishlist(Map<String, dynamic> wishlistData) async {
+    final currentUser = FirebaseAuth.instance.currentUser;
+    if (currentUser == null) {
+      throw Exception('Usuario no autenticado.');
+    }
     try {
-      DocumentReference newWishList = await _db.collection('wishlists').add(wishlistData);
+      DocumentReference newWishList = await _db
+        .collection('users')
+        .doc(currentUser.uid)
+        .collection('wishlists')
+        .add(wishlistData);
       return newWishList.id;
     } catch (e) {
       // Devolvemos el error para que el UI pueda manejarlo
@@ -50,14 +69,20 @@ class WishlistDao {
 
   Future<void> addItem(String wishlistId, Map<String, dynamic> itemData) async {
     try {
+      final currentUser = FirebaseAuth.instance.currentUser;
+      if (currentUser == null) {
+        throw Exception('Usuario no autenticado.');
+      }
       await _db.runTransaction((transaction) async {
-        final wishlistRef = _db.collection('wishlists').doc(wishlistId);
+        final wishlistRef = _db
+          .collection('users')
+          .doc(currentUser.uid)
+          .collection('wishlists').doc(wishlistId);
         CollectionReference itemsRef = wishlistRef.collection('items');
 
         await itemsRef.add(itemData);
         int itemCount = (await itemsRef.count().get()).count ?? 0;
 
-        AggregateQuerySnapshot numItemsSnapshot = await _db.collection('wishlists').doc(wishlistId).collection('items').count().get();
         transaction.update(wishlistRef, {'itemCount': itemCount});
       });
     } catch (e) {
@@ -67,8 +92,19 @@ class WishlistDao {
   }
 
   Future<void> updateItem(String wishlistId, String itemId,  Map<String, dynamic> itemData) async {
+    final currentUser = FirebaseAuth.instance.currentUser;
+    if (currentUser == null) {
+      throw Exception('Usuario no autenticado.');
+    }
     try {
-      await _db.collection('wishlists').doc(wishlistId).collection('items').doc(itemId).set(itemData, SetOptions(merge: true));
+      await _db
+        .collection('users')
+        .doc(currentUser.uid)
+        .collection('wishlists')
+        .doc(wishlistId)
+        .collection('items')
+        .doc(itemId)
+        .set(itemData, SetOptions(merge: true));
     } catch (e) {
       // Devolvemos el error para que el UI pueda manejarlo
       throw Exception('Error al actualizar el deseo: $e');
@@ -76,14 +112,22 @@ class WishlistDao {
   }
 
   Future<void> removeItem(String wishlistId, String itemId) async {
+    final currentUser = FirebaseAuth.instance.currentUser;
+    if (currentUser == null) {
+      throw Exception('Usuario no autenticado.');
+    }
     try {
       await _db.runTransaction((transaction) async {
-        final wishlistRef = _db.collection('wishlists').doc(wishlistId);
+        final wishlistRef = _db
+          .collection('users')
+          .doc(currentUser.uid)
+          .collection('wishlists')
+          .doc(wishlistId);
         CollectionReference itemsRef = wishlistRef.collection('items');
         await itemsRef.doc(itemId).delete();
 
-        AggregateQuerySnapshot numItemsSnapshot = await _db.collection('wishlists').doc(wishlistId).collection('items').count().get();
-        transaction.update(wishlistRef, {'itemCount': (numItemsSnapshot.count??0 - 1)});
+        int itemCount = (await itemsRef.count().get()).count ?? 0;
+        transaction.update(wishlistRef, {'itemCount': itemCount});
       });
     } catch (e) {
       // Devolvemos el error para que el UI pueda manejarlo
@@ -92,19 +136,43 @@ class WishlistDao {
   }
 
   void createOrUpdateWishlist(String id, Map<String, Object> map) {
-     _db.collection('wishlists').doc(id).set(map, SetOptions(merge: true));
+    final currentUser = FirebaseAuth.instance.currentUser;
+    if (currentUser == null) {
+      throw Exception('Usuario no autenticado.');
+    }
+    _db
+     .collection('users')
+      .doc(currentUser.uid)
+      .collection('wishlists')
+      .doc(id)
+      .set(map, SetOptions(merge: true));
   }
 
   void deleteWishlist(String id) {
-     _db.collection('wishlists').doc(id).delete();
+    final currentUser = FirebaseAuth.instance.currentUser;
+    if (currentUser == null) {
+      throw Exception('Usuario no autenticado.');
+    }
+    _db
+      .collection('users')
+      .doc(currentUser.uid)
+      .collection('wishlists')
+      .doc(id)
+      .delete();
   }
 
   Stream<QuerySnapshot<Map<String, dynamic>>> getListItems(WishList currentWishList) {
-    return FirebaseFirestore.instance
-          .collection('wishlists')
-          .doc(currentWishList.id)
-          .collection('items')
-          .snapshots();
+    final currentUser = FirebaseAuth.instance.currentUser;
+    if (currentUser == null) {
+      throw Exception('Usuario no autenticado.');
+    }
+    return _db
+      .collection('users')
+      .doc(currentUser.uid)
+      .collection('wishlists')
+      .doc(currentWishList.id)
+      .collection('items')
+      .snapshots();
   }
 
 }
